ecommerce-management/orderService/src/main/java/com/targetready/orderService/model/Order.java
package com.targetready.orderService.model;


public class Order {
    private String orderId;
    private double amount;

    private String bank;
    private int stock;

    public String getOrderId() {
        return orderId;
    }

    public void setOrderId(String orderId) {
        this.orderId = orderId;
    }

    public double getAmount() {
        return amount;
    }

    public void setAmount(double amount) {
        this.amount = amount;
    }

    public String getBank() {
        return bank;
    }

    public void setBank(String bank) {
        this.bank = bank;
    }

    public int getStock() {
        return stock;
    }

    public void setStock(int stock) {
        this.stock = stock;
    }
}
